cpu_idle_wait handle CPUs in idle
Sometimes cpu_idle_wait gets stuck because it might miss
CPUS that are already in idle, have no tasks waiting to run
and have no interrupts going to them.  This is common on
bootup when switching cpu idle governors.

This patch gives those CPUS that don't check in an IPI kick.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>