rt-time-starvation-fix
Hey Ingo,
	Noticed -rt has been updated a few times and this is still missing so I
figured I'd resend it just in case you missed it:

	We've worked around this before, but its cropped up again. Since
update_wall_time is now called from a softirq, it can be preempted by a
high priority process. If its preempted for long enough, the clocksource
can wrap, causing time to stop incrementing, which if the preempting
process is checking the time, can cause a hard lockup.

This patch forces the clocksource to be read each tick, and accumulate
only the cycle count. This allows the update_wall_time to be deferred
w/o fear of hardware overflow.

thanks
-john

 include/linux/clocksource.h |    2 +-
 include/linux/time.h        |    1 +
 kernel/timer.c              |    1 +
 3 files changed, 3 insertions(+), 1 deletion(-)

linux-2.6.21-rc5_cycles-accumulated_C7.patch
============================================