fix_vdso_gtod_vsyscall64_2
This patch fixes VDSO GTOD enhancements for the case where kernel.vsyscall64=2.
In this case VDSO GTOD presents a cheap way to access gettimeofday() with no
need to issue a real system call. This fix enforces the 1ms resolution VDSO
GTOD should present when kernel.vsyscall64=2.

This patch offers this resolution for the clocksources that have a vread()
function, such as tsc and hpet. Otherwise it keeps uses the read() function
offered by the clocksource, that may be costly.

Also, a few other tweaks are performed.

Signed-off-by: Luis Claudio R. Goncalves <lgoncalv@redhat.com>