sched: lock-break the load balance path
move_tasks() can do a lot of work, and it holds two runqueue locks and has
IRQs disabled. We already introduced sysctl_sched_nr_migrate to limit
the number of task iterations it can do.

This patch takes it one step further and drops the locks once we break out
of the iteration due to sysctl_sched_nr_migrate and re-enables IRQs. Then it
re-acquires everything and continues.

Dropping the locks is safe because:
 - load_balance() doesn't rely on it
 - load_balance_newidle() uses double_lock_balance() which
   can already drop the locks.

Enabling IRQs should be safe since we already dropped all locks.

We add the LB_COMPLETE state to detect the truncated iteration due to
sysctl_sched_nr_migrate.

For now we must break out of the restart when load_moved is 0, because each
iteration will test the same tasks - hence we can live-lock here.

Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>