rt: rwlock fix non nested unlocking
The RW PI locks are sensitive to locking and unlocking order. This is
a bug, since it is not always the case that unlocking will be in the
reverse order of locking.

This patch lets there be "holes" in the rwlock array for a task. The holes
are caused when a lock is released out of order.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>