rwlock: fix modification of owners race
Thomas Gleixner has been hitting a bug where everything would freeze.
It was caused by the race between the owner of a reader lock releasing
it and before it updates its owner count. The pending owner tries to
take the lock, but sees the owners count is too high and fails. This causes
the lock to be unowned, with the owner sleeping.

The fix is to only check the owners count if the lock happens to be held.
Do not check it if the lock is pending.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>