radix-tree: optimistic locking
Implement optimistic locking for the concurrent radix tree.

Optimistic locking is aimed at avoiding taking higher level node locks.
We decent the tree using an RCU lookup, looking for the lowest modification
termination point.

If found, we try to acquire the lock of that node. After we have obtained this
lock, we will need to validate if the initial conditions still hold true. We do
this by repeating the steps that found us this node in the first place.

Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>