ftrace: avoid lockdep recursion
The function_trace_call needs to find the cpu it is running on to be able
to disable the trace to pretect from recursion. It currently performs a
local_irq_save, but when lockdep is on, this may cause unwanted recursion
if lockdep happens to call something that is traced.

This patch changes this to use preempt_disable instead.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>