timers: Use static keys for migrate_enable and nohz_active
The migration_enabled and nohz_active will be later moved into the bitfield. In
the bitfield a change to one bit causes RMW operation and without holding a
lock it might happen that a concurrent change on a second CPU might cause the
loss of the an update.
To avoid that and since both fields are changed to static_branch.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>