timers: Drop expiry lock after each timer invocation
The ->expiry_lock lock is held until every timer is expired. So if a
__del_timer_sync() caller blocks on the lock then it has to wait until
every timer callback has completed.

Therefore drop the lock and acquire it after expiring the timer. To be
able to remove the timer, when it was expired, the running_timer struct
member has to be resetted to NULL as well. This happens after the timer
was expired. It is ok to do this lockless, because the only problem
could be that when a check is done too early, the old expired timer is
stored in there.

Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
[bigeasy: Patch description reworded]
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>