sched/core: Schedule new worker even if PI-blocked
If a task is PI-blocked (blocking on sleeping spinlock) then we don't
schedule a new kworker if we schedule out due to lock contention because
!RT wouldn't do that as well. A spinlock disables preemption and worker
wouldn't schedule out on lock contention (but spin) so it wouldn't start
a new worker.

On RT the RW-semaphore implementation uses an rtmutex so
tsk_is_pi_blocked() will return true if a task blocks on it. In case we
will skip scheduling a worker which may deadlock if one worker is
waiting on progress from another worker. XFS is able to trigger this.

Allow to schedule new worker if the current worker is PI-blocked.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>