softirq: Don't do softirq if ksoftird is pending
!RT does not invoke softirq on return from interrupt if ksoftirqd is
scheduled. This for instance for networking: once the budget is used,
napi will mark NET_RX pending and let ksoftird scheduled. This will
ensure that further napi processing happens in ksoftirqd and the system
can do something else.
In RT we had "shortcut" for this once budget is used to ensure that
further softirq handling happens in ksoftirqd. After the softirq rework
I dropped that shortcut assuming that we would gain mainline behaviour
here. The missing part piece is not to invoke softirqs if ksoftird is
pending.

Don't invoke softirqs on BH-enable if ksoftirqd is scheduled.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>