mm: slub: Move discard_slab() invocations out of IRQ-off sections
discard_slab() gives the memory back to the page-allocator. Some of its
invocation occur from IRQ-disabled sections which were disabled by SLUB.
An example is the deactivate_slab() invocation from within
___slab_alloc() or put_cpu_partial().

Instead of giving the memory back directly, put the pages on a list and
process it once the caller is out of the known IRQ-off region.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>